+++ src/App.tsx
@@ -3,14 +3,14 @@ import { ThemeProvider } from 'styled-components';
 import { theme } from '@styles/theme';
 import Layout from '@components/common/Layout/Layout';
 import Home from '@Pages/Home/Home';
-import Auth from '@Pages/Auth/Auth';
+import Auth from '@Pages/Auth';
 import SignUp from '@Pages/SingUp';
 import Search from '@Pages/Search/Search';
-import Articles from '@Pages/Articles/Articles';
+import Articles from '@Pages/Articles';
 import KakaoRedirectHandler from '@components/Auth/KakaoRedirectHandler';
 import GoogleRedirectHandler from '@components/Auth/GoogleRedirectHandler';
 import NaverRedirectHandler from '@components/Auth/NaverRedirectHandler';
-import WriteComment from '@Pages/WriteComment/WriteComment';
+import WriteComment from '@Pages/WriteComment';
 import Progress2 from '@/Pages/Articles/WriteComment/CreatePost/Progress2/Progress2';
 import Progress3 from './Pages/Articles/WriteComment/CreatePost/Progress3/Progress3';
 import Progress4 from './Pages/Articles/WriteComment/CreatePost/Progress4/Progress4';
