+++ src/Pages/Articles/Articles.tsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import * as A from './ArticlesStyles';
 import CommonHeader from '@components/common/CommonHeader/CommonHeader';
-import WriterProfile from '@/components/Articles/WriterProfile/WriterProfile';
-import ProductInfo from '@components/Articles/ProductInfo/ProductInfo';
-import PostActions from '@components/Articles/PostActions/PostActions';
+import WriterProfile from '@/components/Articles/WriterProfile';
+import ProductInfo from '@components/Articles/ProductInfo';
+import PostActions from '@components/Articles/PostActions';
 import LikeAndComment from '@/components/Articles/LikeAndComment';
 import OfferItemLists from '@/components/Articles/OfferItemLists';
 import useTimeDiff from '@/hooks/useTimeDiff';
