+++ src/Pages/Articles/index.tsx
@@ -1,0 +1,3 @@
+import Articles from './Articles';
+
+export default Articles;
