+++ src/Pages/WriteComment/index.tsx
@@ -1,0 +1,3 @@
+import WriteComment from './WriteComment';
+
+export default WriteComment;
