+++ src/components/Articles/PostActions/index.tsx
@@ -1,0 +1,3 @@
+import PostActions from './PostActions';
+
+export default PostActions;
