+++ src/components/Articles/ProductInfo/index.tsx
@@ -1,0 +1,3 @@
+import ProductInfo from './ProductInfo';
+
+export default ProductInfo;
