+++ src/components/Articles/WriterProfile/index.tsx
@@ -1,0 +1,3 @@
+import WriterProfile from './WriterProfile';
+
+export default WriterProfile;
