+++ src/components/WriteComment/ContentsSection/ContentsSection.tsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 0 21px;
+  margin-top: 77px;
+`;
+
+interface IContentsSectionProps {
+  children: React.ReactNode;
+}
+
+function ContentsSection({ children }: IContentsSectionProps) {
+  return <Container>{children}</Container>;
+}
+
+export default ContentsSection;
