+++ src/components/WriteComment/ContentsSection/index.tsx
@@ -1,18 +1,3 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 0 21px;
-  margin-top: 77px;
-`;
-
-interface IContentsSectionProps {
-  children: React.ReactNode;
-}
-
-function ContentsSection({ children }: IContentsSectionProps) {
-  return <Container>{children}</Container>;
-}
+import ContentsSection from './ContentsSection';
 
 export default ContentsSection;
