+++ src/components/common/ProfileSetupForm/CurrentLocation/index.tsx
@@ -1,27 +1,3 @@
-import * as C from './CurrentLocationStyles';
-import BigTitle from '@components/common/BigTitle';
-import BlueButton from '@/components/common/Buttons/BlueButton';
-
-interface ICurrentLocationProps {
-  addressInfo: string;
-  handleCompleteBtn: () => void;
-}
-
-function CurrentLocation({
-  addressInfo,
-  handleCompleteBtn,
-}: ICurrentLocationProps) {
-  return (
-    <C.Wrapper>
-      <C.TitleContainer>
-        <C.SmallTitle>현재 동네</C.SmallTitle>
-        <BigTitle>{addressInfo}</BigTitle>
-      </C.TitleContainer>
-      <BlueButton disabled={false} onClick={handleCompleteBtn} maxWidth="100%">
-        이 동네로 설정
-      </BlueButton>
-    </C.Wrapper>
-  );
-}
+import CurrentLocation from './CurrentLocation';
 
 export default CurrentLocation;
